requirements.txt
fastapi
uvicorn
langchain
langchain-chroma
langchain-openai
langchain-huggingface
chromadb
python-dotenv
gradio
pysqlite3-binary